HV-461 Moving test util classes into a new package named testutil.
