+++ hibernate-validator/src/test/java/org/hibernate/validator/test/engine/ValidatorTest.java
@@ -31,12 +31,12 @@ import javax.validation.metadata.BeanDescriptor;
 import org.testng.annotations.Test;
 
 import org.hibernate.validator.constraints.Length;
-import org.hibernate.validator.test.util.CountValidationCalls;
-import org.hibernate.validator.test.util.CountValidationCallsValidator;
+import org.hibernate.validator.test.testutil.CountValidationCalls;
+import org.hibernate.validator.test.testutil.CountValidationCallsValidator;
 
-import static org.hibernate.validator.test.util.ConstraintViolationAssert.assertCorrectPropertyPaths;
-import static org.hibernate.validator.test.util.ConstraintViolationAssert.assertNumberOfViolations;
-import static org.hibernate.validator.test.util.ValidatorUtil.getValidator;
+import static org.hibernate.validator.test.testutil.ConstraintViolationAssert.assertCorrectPropertyPaths;
+import static org.hibernate.validator.test.testutil.ConstraintViolationAssert.assertNumberOfViolations;
+import static org.hibernate.validator.test.testutil.ValidatorUtil.getValidator;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
