+++ hibernate-validator/src/test/java/org/hibernate/validator/test/engine/failfast/FailFastTest.java
@@ -22,12 +22,12 @@ import org.hibernate.validator.constraints.Email;
 import org.hibernate.validator.constraints.NotBlank;
 import org.hibernate.validator.method.MethodConstraintViolationException;
 import org.hibernate.validator.method.MethodValidator;
-import org.hibernate.validator.test.util.ValidatorUtil;
-import org.hibernate.validator.test.util.ValidationInvocationHandler;
+import org.hibernate.validator.test.testutil.ValidationInvocationHandler;
+import org.hibernate.validator.test.testutil.ValidatorUtil;
 import org.hibernate.validator.util.LoggerFactory;
 
-import static org.hibernate.validator.test.util.ConstraintViolationAssert.assertNumberOfViolations;
-import static org.hibernate.validator.test.util.ValidatorUtil.getMethodValidationProxy;
+import static org.hibernate.validator.test.testutil.ConstraintViolationAssert.assertNumberOfViolations;
+import static org.hibernate.validator.test.testutil.ValidatorUtil.getMethodValidationProxy;
 import static org.testng.Assert.fail;
 
 /**
