+++ hibernate-validator/src/test/java/org/hibernate/validator/test/engine/groups/defaultgroupsequenceprovider/DefaultGroupSequenceProviderTest.java
@@ -35,9 +35,9 @@ import org.hibernate.validator.group.GroupSequenceProvider;
 import org.hibernate.validator.method.MethodConstraintViolation;
 import org.hibernate.validator.method.MethodValidator;
 
-import static org.hibernate.validator.test.util.ConstraintViolationAssert.assertCorrectConstraintViolationMessages;
-import static org.hibernate.validator.test.util.ConstraintViolationAssert.assertNumberOfViolations;
-import static org.hibernate.validator.test.util.ValidatorUtil.getValidator;
+import static org.hibernate.validator.test.testutil.ConstraintViolationAssert.assertCorrectConstraintViolationMessages;
+import static org.hibernate.validator.test.testutil.ConstraintViolationAssert.assertNumberOfViolations;
+import static org.hibernate.validator.test.testutil.ValidatorUtil.getValidator;
 
 /**
  * @author Kevin Pollet - SERLI - (kevin.pollet@serli.com)
