+++ hibernate-validator/src/test/java/org/hibernate/validator/test/engine/methodlevel/IllegalMethodParameterConstraintsTest.java
@@ -28,9 +28,9 @@ import org.testng.annotations.Test;
 import org.hibernate.validator.engine.ValidatorImpl;
 import org.hibernate.validator.metadata.BeanMetaDataImpl;
 
-import static org.hibernate.validator.test.util.ConstraintViolationAssert.assertCorrectConstraintViolationMessages;
-import static org.hibernate.validator.test.util.ValidatorUtil.getMethodValidator;
-import static org.hibernate.validator.test.util.ValidatorUtil.getValidator;
+import static org.hibernate.validator.test.testutil.ConstraintViolationAssert.assertCorrectConstraintViolationMessages;
+import static org.hibernate.validator.test.testutil.ValidatorUtil.getMethodValidator;
+import static org.hibernate.validator.test.testutil.ValidatorUtil.getValidator;
 
 /**
  * Integration test for {@link ValidatorImpl} and {@link BeanMetaDataImpl} which
