+++ hibernate-validator/src/test/java/org/hibernate/validator/test/engine/proxy/ProxyTest.java
@@ -26,8 +26,9 @@ import javax.validation.Validator;
 import static org.testng.Assert.assertEquals;
 import org.testng.annotations.Test;
 
-import org.hibernate.validator.test.util.ValidatorUtil;
-import static org.hibernate.validator.test.util.ConstraintViolationAssert.assertNumberOfViolations;
+import org.hibernate.validator.test.testutil.ValidatorUtil;
+
+import static org.hibernate.validator.test.testutil.ConstraintViolationAssert.assertNumberOfViolations;
 
 /**
  * See HV-257
