+++ hibernate-validator/src/test/java/org/hibernate/validator/test/metadata/BeanMetaDataImplTest.java
@@ -31,7 +31,7 @@ import org.hibernate.validator.metadata.ConstraintHelper;
 import org.hibernate.validator.test.engine.methodlevel.service.CustomerRepository;
 import org.hibernate.validator.test.engine.methodlevel.service.CustomerRepositoryImpl;
 
-import static org.hibernate.validator.test.util.ConstraintViolationAssert.assertIterableSize;
+import static org.hibernate.validator.test.testutil.ConstraintViolationAssert.assertIterableSize;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
