+++ hibernate-validator/src/test/java/org/hibernate/validator/test/metadata/MethodDescriptorTest.java
@@ -30,7 +30,7 @@ import org.hibernate.validator.method.metadata.ParameterDescriptor;
 import org.hibernate.validator.test.metadata.CustomerRepository.ValidationGroup;
 import org.hibernate.validator.test.metadata.CustomerRepositoryExt.CustomerExtension;
 
-import static org.hibernate.validator.test.util.ValidatorUtil.getMethodDescriptor;
+import static org.hibernate.validator.test.testutil.ValidatorUtil.getMethodDescriptor;
 import static org.hibernate.validator.util.Contracts.assertNotNull;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
